pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "equation-surface-modeller"
version = "0.1.0"
description = "Program for modelling 3d surfaces with equations"
readme = "README.md"
authors = [{name = "Igor Razin"}]
license = {text = "MIPT"}
requires-python = ">=3.8"

dependencies = [
    "numpy>=1.21.0",
    "cexprtk>=0.4.2",
    "pyvista>=0.47.0",
    "PyMCubes>=0.1.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["src/calc/ut"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
